added canvas text test, cleanup files, move text render tests to seperate folder
